feat(app): logged-out console with locked features
A signed-out visitor arriving from the marketing site used to get a
chrome-free card at "/" — amount, payer count, address, "Get link", and a
"Log in" text link. They never saw the product they were being sold.

Now "/" drops them into the real console: sidebar, topbar, mobile nav, and
the full create-request screen. Creating a request still needs no account
(POST /api/requests is session-optional). Everything that DOES need one —
the requests list, approvals, treasury, people, groups, settings, the
workspace switcher, group-split and contract-linking — renders as a real,
disabled control with an inline "Sign in to <do the thing>" reason.

Making the shell safe to mount signed out needed two things:

  - every session-requiring query gated (OnboardingGate's useGetUser was
    ungated and 401'd on every anonymous page load; the locked routes now
    gate by never mounting the component that owns the query)
  - the 401 interceptor taught to tell an anonymous visitor apart from an
    expired session, via session-cookie presence at server render
    (lib/session-presence.ts). No cookie ever => never redirect. Cookie
    present => a real expiry, still hard-redirects to /login.

The create screen moved to components/create/ because "/" and "/create"
now mount the same thing, and it gained a payer-count stepper: the rich
form hardcoded payerCount 1, so folding the old landing page into it would
have silently dropped the anonymous multi-payer capability.
