+++ app/requests/[id]/page.tsx
@@ -16,6 +16,8 @@ import {
 import { markParticipantAsPaid } from "@/features/expenses/api/client";
 import { sendReminder } from "@/features/reminders/api/client";
 import { usePageTitle } from "@/contexts/page-title";
+import { useAuthStore } from "@/stores/authStore";
+import { LockedScreen } from "@/components/shell/locked-feature";
 import {
   StatusChip,
   STATUS_STYLE,
@@ -330,7 +332,7 @@ function PayerRow({
   );
 }
 
-export default function RequestDetailPage() {
+function RequestDetailScreen() {
   const params = useParams();
   const router = useRouter();
   const id = params?.id as string;
@@ -652,4 +654,26 @@ function NudgeRow({
       </Btn>
     </div>
   );
+}
+
+/**
+ * The shell now renders its chrome for signed-out visitors too, so
+ * `/requests/[id]` — including any real request id typed or bookmarked — is
+ * reachable without a session. Gate here, above RequestDetailScreen's
+ * `getRequestDetail` fetch: hooks can't be called conditionally, so the only
+ * way to keep that request-detail lookup from firing for an anonymous
+ * visitor is to never mount the component that owns it.
+ */
+export default function RequestDetailPage() {
+  const { isAuthenticated } = useAuthStore();
+  if (!isAuthenticated) {
+    return (
+      <LockedScreen
+        title="Request details"
+        reason="Sign in to see this request"
+        blurb="Tracking a request — its payers, its status and its history — belongs to the account that created it."
+      />
+    );
+  }
+  return <RequestDetailScreen />;
 }
