+++ app/treasury/page.tsx
@@ -10,6 +10,8 @@ import { formatCurrency } from "@/utils/formatters";
 import { LogIncomeModal } from "@/components/log-income-modal";
 import { Row } from "@/components/shell/row";
 import { Card, HeroCard, Eyebrow, Btn, Icons, T, TYPE, getUserColor, G as G_ACCENT } from "@/lib/splito-design";
+import { useAuthStore } from "@/stores/authStore";
+import { LockedScreen } from "@/components/shell/locked-feature";
 
 /** "Received 1 Aug" — the trailing date on a stream row (design line 1229). */
 function formatReceivedDate(date: Date): string {
@@ -24,7 +26,7 @@ function formatReceivedDate(date: Date): string {
  * hero total instead reads `workspaces/:id/summary`'s `treasury` field so it
  * always agrees with the dashboard rather than being summed here separately.
  */
-export default function TreasuryPage() {
+function TreasuryScreen() {
   const workspace = useActiveWorkspace();
   const isResolving = useIsResolvingWorkspace();
   const isBusiness = workspace.kind === "business";
@@ -161,4 +163,27 @@ export default function TreasuryPage() {
       />
     </div>
   );
+}
+
+/**
+ * The shell now renders its chrome for signed-out visitors too, so
+ * `/treasury` is reachable without a session. Gate here, above
+ * TreasuryScreen's `useGetStreamsByOrganization`/`useWorkspaceTreasury`
+ * hooks: those pass `enabled: isBusiness` to skip the network call, but the
+ * hooks themselves still mount for an anonymous visitor with a workspace,
+ * and hooks can't be called conditionally — so the only way to stop them is
+ * to never mount the component that owns them.
+ */
+export default function TreasuryPage() {
+  const { isAuthenticated } = useAuthStore();
+  if (!isAuthenticated) {
+    return (
+      <LockedScreen
+        title="Treasury"
+        reason="Sign in to see your treasury"
+        blurb="What has come in, where it came from, and what it settled into."
+      />
+    );
+  }
+  return <TreasuryScreen />;
 }
